Add similarity and AP tests
